+++ packages/desktop-client/package.json
@@ -64,8 +64,8 @@
     "#components/reports/numberFormatter": "./src/components/reports/numberFormatter.ts",
     "#components/reports/ReportOptions": "./src/components/reports/ReportOptions.ts",
     "#components/reports/reportRanges": "./src/components/reports/reportRanges.ts",
-    "#components/reports/reports/monteCarloHistoricalReturns": "./src/components/reports/reports/monteCarloHistoricalReturns.ts",
-    "#components/reports/reports/monteCarloSimulation": "./src/components/reports/reports/monteCarloSimulation.ts",
+    "#components/reports/reports/monte-carlo/monteCarloHistoricalReturns": "./src/components/reports/reports/monte-carlo/monteCarloHistoricalReturns.ts",
+    "#components/reports/reports/monte-carlo/monteCarloSimulation": "./src/components/reports/reports/monte-carlo/monteCarloSimulation.ts",
     "#components/reports/setSessionReport": "./src/components/reports/setSessionReport.ts",
     "#components/reports/spendingAverageRange": "./src/components/reports/spendingAverageRange.ts",
     "#components/reports/spreadsheets/cash-flow-spreadsheet": "./src/components/reports/spreadsheets/cash-flow-spreadsheet.tsx",
